# api/docs/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: AstraPBX API
  description: |
    AstraPBX is a multi-tenant cloud PBX platform built on Asterisk. This API provides complete control over telephony infrastructure — from provisioning SIP trunks and DIDs to managing call queues, users, greetings, and hold music.

    ## Authentication

    AstraPBX supports two authentication methods. Set your credentials once using the **Authorize** button above — they persist across all requests in this session.

    ### Option 1: API Key (Recommended)
    Pass your organization's API key and secret as headers. No login step required.
    ```
    X-API-Key: org_xxxxxxxxxxxx
    X-API-Secret: your-api-secret
    ```

    ### Option 2: Bearer Token (JWT)
    Authenticate via `POST /auth/login` with your API key and secret to receive a JWT token. Pass it as a Bearer token in subsequent requests.
    ```
    Authorization: Bearer eyJhbGciOi...
    ```

    ### Admin Endpoints
    Endpoints under `/admin/*` require admin credentials and are used only for initial organization setup. Do not use admin tokens for regular API operations.

    ## Auto-Deployment

    Most write operations (creating users, updating queues, changing routing) automatically deploy configuration changes to Asterisk and reload. No manual deployment step is needed unless you want to force a full redeploy via `/config/deploy`.

    ## Quick Start

    1. **Authenticate** — Use the Authorize button with your API Key + Secret
...

  version: 1.0.0
...
servers:
...
security:
  - ApiKeyAuth: []
    ApiSecretAuth: []
  - BearerAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Organization API key (e.g., `org_<32-char-hex>`)
    ApiSecretAuth:
      type: apiKey
      in: header
      name: X-API-Secret
      description: Organization API secret
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from `POST /auth/login`
  schemas:
...
    CallRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        call_id:
          type: string
        from_number:
          type: string
        to_number:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [ringing, answered, busy, failed, no-answer, completed]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        duration:
          type: integer
        recording_url:
          type: string
...
paths:
...
  /calls/count:
    get:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Get call count and statistics
      description: |
        Retrieve call statistics with optional filters

        **Example curl:**
        ```bash
        curl -X GET "${YOUR_SERVER_URL}/api/v1/calls/count?status=completed&from=2025-01-01T00:00:00Z&to=2025-12-31T23:59:59Z" \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, failed]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Call statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  active:
                    type: integer
                  completed:
                    type: integer
                  failed:
                    type: integer
                  average_duration:
                    type: integer
                  total_duration:
                    type: integer
                example:
                  total: 150
                  active: 3
                  completed: 140
                  failed: 7
                  average_duration: 245
                  total_duration: 34300
...
  /calls/channels:
    get:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Get active channels with full details
      description: |
        Retrieve all active Asterisk channels for this organization with detailed channel information.
        Returns channel IDs needed for call control operations (hangup, hold, transfer).

        **Use this endpoint to get `channel_id` values required by:**
        - `POST /calls/{channelId}/hangup`
        - `POST /calls/{channelId}/hold`
        - `POST /calls/{channelId}/unhold`
        - `POST /calls/{channelId}/transfer`
      responses:
        '200':
          description: Active channels
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  org_id:
                    type: string
                    format: uuid
                  channels:
                    type: array
                    items:
                      type: object
                      properties:
                        channel_id:
                          type: string
                          example: "PJSIP/1001-00000003"
                        channel_name:
                          type: string
                        uniqueid:
                          type: string
                        linkedid:
                          type: string
                        state:
                          type: string
                          enum: [Up, Ringing, Ring, Down, Busy]
                        caller_number:
                          type: string
                        caller_name:
                          type: string
                        connected_number:
                          type: string
                        connected_name:
                          type: string
                        context:
                          type: string
                        extension:
                          type: string
                        application:
                          type: string
                        application_data:
                          type: string
                        duration:
                          type: integer
                        bridge_id:
                          type: string
  /calls/{channelId}/hangup:
    post:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Hang up an active call
      description: |
        Terminate an active call immediately. Get the `channelId` from `GET /calls/channels` or `GET /calls/live`.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: Asterisk channel ID (e.g., PJSIP/1001-00000003)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [normal, busy, congestion]
                  default: normal
                  description: Hangup cause
      responses:
        '200':
          description: Call terminated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  channel_id:
                    type: string
                  reason:
                    type: string
  /calls/{channelId}/hold:
    post:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Put a call on hold
      description: |
        Place an active call on hold. The caller hears music on hold until the call is resumed via `POST /calls/{channelId}/unhold`.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: Asterisk channel ID
      responses:
        '200':
          description: Call placed on hold
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  channel_id:
                    type: string
  /calls/{channelId}/unhold:
    post:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Resume a call from hold
      description: |
        Remove a call from hold and redirect it to a destination extension.
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: Asterisk channel ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination]
              properties:
                destination:
                  type: string
                  description: Extension to redirect the call to after unhold
                  example: "1001"
      responses:
        '200':
          description: Call resumed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  channel_id:
                    type: string
                  destination:
                    type: string
...
  /calls/{channelId}/transfer:
    post:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Transfer an active call
      description: |
        Transfer an active call to another destination (extension, queue, etc.).
        This performs a blind transfer using Asterisk ARI.

        **Example curl:**
        ```bash
        curl -X POST ${YOUR_SERVER_URL}/api/v1/calls/CHANNEL_ID/transfer \
          -H "Authorization: Bearer YOUR_ORG_TOKEN" \
          -H "Content-Type: application/json" \
          -d '{
            "destination": "2003",
            "type": "blind",
            "context": "testorg_internal"
          }'
        ```
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
          description: Asterisk channel ID (e.g., PJSIP/2001-00000001)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination]
              properties:
                destination:
                  type: string
                  description: Transfer destination (extension, queue number, etc.)
                  example: "2003"
                type:
                  type: string
                  enum: [blind, attended]
                  default: blind
                  description: Transfer type
                context:
                  type: string
                  description: Asterisk context (optional, auto-determined from org)
                  example: "testorg_internal"
      responses:
        '200':
          description: Call transferred successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Call transferred successfully"
                  transfer:
                    type: object
                    properties:
                      channel_id:
                        type: string
                      destination:
                        type: string
                      context:
                        type: string
                      type:
                        type: string
        '400':
          description: Bad request - missing destination
        '401':
          description: Unauthorized - Invalid or missing token
        '503':
          description: Service unavailable - ARI not connected
        '500':
          description: Server error - Transfer failed
  /calls:
    get:
      security:
        - BearerAuth: []
      tags: [Call Management]
      summary: Get call history
      description: |
        Retrieve call history with pagination and filtering.

        **Example curl:**
        ```bash
        curl -X GET "${YOUR_SERVER_URL}/api/v1/calls?limit=50&offset=0&status=answered" \
          -H "Authorization: Bearer YOUR_ORG_TOKEN"
        ```
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of records to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of records to skip
        - name: status
          in: query
          schema:
            type: string
            enum: [ringing, answered, busy, failed, no-answer, completed]
          description: Filter by call status
        - name: direction
          in: query
          schema:
            type: string
            enum: [inbound, outbound, internal]
          description: Filter by call direction
      responses:
        '200':
          description: Call history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items:
                      $ref: '#/components/schemas/CallRecord'
                  count:
                    type: integer
                    description: Number of calls in this response
                  total:
                    type: integer
                    description: Total number of calls matching filters
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized
...